chore(audio/workstation): added stereo-to-mono sink
Every now and then there is a video which has messed up audio channels.
E.g: everything is on the left channel, alternatively voice in left and
main audio in right.
This captures and combines both channels and outputs it as a mono signal

Yoinked from: https://askubuntu.com/a/1451226