chore(hm-edit-file): added hyprctl reload for when hyprland.conf is edited
Before 0.47.2, Hyprland happily switched to the dereferenced file after
running this helper.
After 0.47.2, Hyprland reverts to default config when the symlink is
dereferenced and I have to either close this helper to open another
terminal or remember to open one prior to running it.

Could be related to https://github.com/hyprwm/Hyprland/pull/9219 but I
don't have time to check just yet.